% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_drug_record}
\alias{get_drug_record}
\title{for a given record from daily med search, return drug properties
Note: Use query_daily_med first!}
\usage{
get_drug_record(record, namespaces = c(n = "urn:hl7-org:v3"))
}
\arguments{
\item{record}{drug record to search}

\item{namespaces}{xml namespace to use for parsing}
}
\description{
for a given record from daily med search, return drug properties
Note: Use query_daily_med first!
}
