% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query_daily_med}
\alias{query_daily_med}
\title{Query drugname using Daily Med API}
\usage{
query_daily_med(query, by = "drugname", ...)
}
\arguments{
\item{query}{string to query}

\item{by}{search by drugnames or drug classes}

\item{...}{other parameters}
}
\description{
Query drugname using Daily Med API
}
